+++ ChangeLog.md
@@ -4,9 +4,12 @@ ___
 
 ## Unreleased
 
-- NEW : **CONTROLLER SYSTEM** - *20/07/2021* - 1.15.0
+- NEW : **CONTROLLER SYSTEM** - *20/07/2021* - 1.16.0
   The controller system was missing but now we can use it and make very cool stuffs
-- FIX - Compatibility V14 : Edit the descriptor: $this->family - *2021-06-10* - 1.14.4
+
+- NEW : support supplier invoices - *2021-06-16* - 1.15.0
+- FIX : Add contact to ticket on creation - *2021-06-20* - 1.14.4
+- FIX : V14 compatibility : Edit the descriptor: $this->family - *2021-06-10* - 1.14.4
 - NEW : Option to set favicon
 
 ## RELEASE 1.13 - 2021-03-19